+++ include/Ark/Compiler/Macros/Executors/Pipeline.hpp
@@ -1,19 +1,20 @@
 /**
- * @file MacroExecutorPipeline.hpp
- * @author Ray John Alovera (rakista112@gmail.com)
+ * @file Pipeline.hpp
+ * @author Ray John Alovera (rakista112@gmail.com), Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief The Chain of Responsibility class for running nodes through MacroExecutors
- * @version 0.2
+ * @version 0.3
  * @date 2021-05-04
  * 
  * @copyright Copyright (c) 2021
  * 
  */
 
-#ifndef ARK_COMPILER_EXECUTORPIPELINE_HPP
-#define ARK_COMPILER_EXECUTORPIPELINE_HPP
+#ifndef MACROS_EXECUTORS_PIPELINE_HPP
+#define MACROS_EXECUTORS_PIPELINE_HPP
 
 #include <vector>
-#include <Ark/Compiler/MacroExecutor.hpp>
+
+#include <Ark/Compiler/Macros/Executor.hpp>
 
 namespace Ark::internal
 {
@@ -26,6 +27,8 @@ namespace Ark::internal
     class MacroExecutorPipeline
     {
     public:
+        MacroExecutorPipeline() = default;
+
         /**
          * @brief Construct a new Macro Executor Pipeline object
          *